media/diagrams/requirements.txt
daft-pgm
matplotlib